Some methods moved to JuniperBaseHandler
